forge fmt

+++ src/routers/Gateway.sol
@@ -16,7 +16,13 @@ interface ConnectorLike {
     ) external;
     function updateMember(uint64 poolId, bytes16 trancheId, address user, uint64 validUntil) external;
     function updateTokenPrice(uint64 poolId, bytes16 trancheId, uint128 price) external;
-    function handleTransfer(uint64 poolId, bytes16 trancheId,uint256 destinationChainId, address destinationAddress, uint128 amount) external;
+    function handleTransfer(
+        uint64 poolId,
+        bytes16 trancheId,
+        uint256 destinationChainId,
+        address destinationAddress,
+        uint128 amount
+    ) external;
 }
 
 interface RouterLike {
@@ -75,15 +81,18 @@ contract ConnectorGateway {
     }
 
     // --- Outgoing ---
-    function transferToCentrifuge(
-        uint64 poolId,
-        bytes16 trancheId,
-        bytes32 destinationAddress,
-        uint128 amount
-    ) public onlyConnector {
+    function transferToCentrifuge(uint64 poolId, bytes16 trancheId, bytes32 destinationAddress, uint128 amount)
+        public
+        onlyConnector
+    {
         router.send(
             ConnectorMessages.formatTransfer(
-                poolId, trancheId, ConnectorMessages.formatDomain(ConnectorMessages.Domain.Centrifuge), uint(0), destinationAddress, amount
+                poolId,
+                trancheId,
+                ConnectorMessages.formatDomain(ConnectorMessages.Domain.Centrifuge),
+                uint256(0),
+                destinationAddress,
+                amount
             )
         );
     }
@@ -97,7 +106,12 @@ contract ConnectorGateway {
     ) public onlyConnector {
         router.send(
             ConnectorMessages.formatTransfer(
-                poolId, trancheId, ConnectorMessages.formatDomain(ConnectorMessages.Domain.EVM), destinationChainId, destinationAddress, amount
+                poolId,
+                trancheId,
+                ConnectorMessages.formatDomain(ConnectorMessages.Domain.EVM),
+                destinationChainId,
+                destinationAddress,
+                amount
             )
         );
     }
@@ -120,8 +134,8 @@ contract ConnectorGateway {
             (uint64 poolId, bytes16 trancheId, uint128 price) = ConnectorMessages.parseUpdateTokenPrice(_msg);
             connector.updateTokenPrice(poolId, trancheId, price);
         } else if (ConnectorMessages.isTransfer(_msg)) {
-            (uint64 poolId, bytes16 trancheId,,uint256 destinationChainId, address destinationAddress, uint128 amount) =
-                ConnectorMessages.parseTransfer20(_msg);
+            (uint64 poolId, bytes16 trancheId,, uint256 destinationChainId, address destinationAddress, uint128 amount)
+            = ConnectorMessages.parseTransfer20(_msg);
             connector.handleTransfer(poolId, trancheId, destinationChainId, destinationAddress, amount);
         } else {
             revert("ConnectorGateway/invalid-message");
